+++ src/test/java/org/fross/rpncalc/StackConstantsTest.java
@@ -44,7 +44,7 @@ class StackConstantsTest {
 		assertEquals(1, calcStack.size());
 
 		// Ensure that the value is correct
-		assertEquals(java.lang.Math.PI, calcStack.get(0));
+		assertEquals(java.lang.Math.PI, calcStack.get(0).doubleValue());
 	}
 
 	@Test
@@ -59,7 +59,7 @@ class StackConstantsTest {
 		assertEquals(1, calcStack.size());
 
 		// Ensure that the value is correct
-		assertEquals(1.61803398874989, calcStack.get(0));
+		assertEquals(1.61803398874989, calcStack.get(0).doubleValue());
 	}
 
 	@Test
@@ -74,7 +74,7 @@ class StackConstantsTest {
 		assertEquals(1, calcStack.size());
 
 		// Ensure that the value is correct
-		assertEquals(2.7182818284590452353602874713527, calcStack.get(0));
+		assertEquals(2.7182818284590452353602874713527, calcStack.get(0).doubleValue());
 	}
 
 	@Test
@@ -89,7 +89,7 @@ class StackConstantsTest {
 		assertEquals(1, calcStack.size());
 
 		// Ensure that the value is correct
-		assertEquals((double) 299792458, calcStack.get(0));
+		assertEquals((double) 299792458, calcStack.get(0).doubleValue());
 	}
 
 }
